[C] Move requests hash to API module
