+++ client/src/api/index.js
@@ -1,18 +1,22 @@
+export annotationsAPI from 'api/annotations';
+export ApiClient from 'api/client';
+export collectionsAPI from 'api/collections';
+export favoriteProjectsAPI from 'api/favoriteProjects';
+export favoritesAPI from 'api/favorites';
+export makersAPI from 'api/makers';
+export meAPI from 'api/me';
+export pagesAPI from 'api/pages';
+export passwordsAPI from 'api/passwords';
 export projectsAPI from 'api/projects';
 export resourcesAPI from 'api/resources';
+export sectionsAPI from 'api/sections';
+export settingsAPI from 'api/settings';
+export statisticsAPI from 'api/statistics';
 export subjectsAPI from 'api/subjects';
+export textCategoriesAPI from 'api/textCategories';
 export textsAPI from 'api/texts';
 export tokensAPI from 'api/tokens';
 export usersAPI from 'api/users';
-export sectionsAPI from 'api/sections';
-export meAPI from 'api/me';
-export favoritesAPI from 'api/favorites';
-export favoriteProjectsAPI from 'api/favoriteProjects';
-export pagesAPI from 'api/pages';
-export collectionsAPI from 'api/collections';
-export makersAPI from 'api/makers';
-export settingsAPI from 'api/settings';
-export textCategoriesAPI from 'api/textCategories';
-export passwordsAPI from 'api/passwords';
-export ApiClient from 'api/client';
+
+export requests from 'api/requests';
