+++ client/src/containers/backend/People/Makers/Edit.js
@@ -3,14 +3,14 @@ import { connect } from 'react-redux';
 import { Drawer, Dialog } from 'components/backend';
 import { entityStoreActions } from 'actions';
 import { entityUtils } from 'utils';
-import { makersAPI } from 'api';
-const { select } = entityUtils;
+import { makersAPI, requests } from 'api';
 import { Form } from 'components/backend';
 import { Form as FormContainer } from 'containers/backend';
 import { browserHistory } from 'react-router';
 import get from 'lodash/get';
 
-const { request, flush, requests } = entityStoreActions;
+const { request, flush } = entityStoreActions;
+const { select } = entityUtils;
 
 class MakersEditContainer extends PureComponent {
