+++ client/src/containers/backend/ProjectDetail/Collaborators.js
@@ -1,13 +1,13 @@
 import React, { Component, PropTypes } from 'react';
 import { Form, Text } from 'components/backend';
-import { projectsAPI } from 'api';
+import { projectsAPI, requests } from 'api';
 import { Form as FormContainer } from 'containers/backend';
 import { connect } from 'react-redux';
 import { entityStoreActions } from 'actions';
 import { entityUtils } from 'utils';
 import get from 'lodash/get';
 
-const { request, flush, requests } = entityStoreActions;
+const { request, flush } = entityStoreActions;
 
 class ProjectDetailCollaborators extends Component {
