+++ client/src/containers/backend/Settings/General.js
@@ -2,10 +2,8 @@ import React, { PureComponent, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import { Form, MaskedTextInput } from 'components/backend';
 import { Form as FormContainer } from 'containers/backend';
-import { settingsAPI } from 'api';
-import { entityStoreActions } from 'actions';
+import { settingsAPI, requests } from 'api';
 import { entityUtils } from 'utils';
-const { requests } = entityStoreActions;
 
 class SettingsGeneralContainer extends PureComponent {
